Refactor PodSpec decoration and JobSpec API packages
The `pjutil` package is growing bloated and starting to accrue features
that are not entirely coherent. Furthermore, future work will need to be
able to import parts of the package independently in order to not create
cyclic dependencies.
