+++ prow/cmd/clonerefs/BUILD.bazel
@@ -9,8 +9,8 @@ go_library(
     deps = [
         "//prow/kube:go_default_library",
         "//prow/logrusutil:go_default_library",
-        "//prow/pjutil:go_default_library",
         "//prow/pod-utils/clone:go_default_library",
+        "//prow/pod-utils/downwardapi:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/sets:go_default_library",
     ],
