+++ prow/cmd/tot/BUILD.bazel
@@ -36,6 +36,7 @@ go_library(
         "//prow/config:go_default_library",
         "//prow/logrusutil:go_default_library",
         "//prow/pjutil:go_default_library",
+        "//prow/pod-utils/downwardapi:go_default_library",
         "//prow/pod-utils/gcs:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
     ],
