+++ prow/cmd/tot/fallbackcheck/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
     deps = [
         "//prow/config:go_default_library",
         "//prow/pjutil:go_default_library",
+        "//prow/pod-utils/downwardapi:go_default_library",
         "//prow/pod-utils/gcs:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
