+++ prow/pjutil/BUILD.bazel
@@ -11,8 +11,8 @@ load(
 go_library(
     name = "go_default_library",
     srcs = [
-        "jobspec.go",
         "pjutil.go",
+        "tot.go",
     ],
     importpath = "k8s.io/test-infra/prow/pjutil",
     tags = ["automanaged"],
@@ -20,9 +20,9 @@ go_library(
         "//prow/config:go_default_library",
         "//prow/github:go_default_library",
         "//prow/kube:go_default_library",
+        "//prow/pod-utils/downwardapi:go_default_library",
         "//vendor/github.com/satori/go.uuid:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
-        "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
     ],
 )
@@ -42,17 +42,11 @@ filegroup(
 
 go_test(
     name = "go_default_test",
-    srcs = [
-        "jobspec_test.go",
-        "pjutil_test.go",
-    ],
+    srcs = ["pjutil_test.go"],
     embed = [":go_default_library"],
     tags = ["automanaged"],
     deps = [
         "//prow/kube:go_default_library",
-        "//vendor/k8s.io/api/core/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/diff:go_default_library",
     ],
 )
