+++ prow/pjutil/pjutil.go
@@ -18,12 +18,8 @@ limitations under the License.
 package pjutil
 
 import (
-	"fmt"
-	"sort"
-
 	"github.com/satori/go.uuid"
 	"github.com/sirupsen/logrus"
-	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"k8s.io/test-infra/prow/config"
@@ -136,70 +132,6 @@ func BatchSpec(p config.Presubmit, refs kube.Refs) kube.ProwJobSpec {
 		pjs.RunAfterSuccess = append(pjs.RunAfterSuccess, BatchSpec(nextP, refs))
 	}
 	return pjs
-}
-
-// ProwJobToPod converts a ProwJob to a Pod that will run the tests.
-func ProwJobToPod(pj kube.ProwJob, buildID string) (*v1.Pod, error) {
-	if pj.Spec.PodSpec == nil {
-		return nil, fmt.Errorf("prowjob %q lacks a pod spec?", pj.Name)
-	}
-	env, err := EnvForSpec(NewJobSpec(pj.Spec, buildID, pj.Name))
-	if err != nil {
-		return nil, err
-	}
-
-	spec := pj.Spec.PodSpec.DeepCopy()
-	spec.RestartPolicy = "Never"
-
-	for i := range spec.InitContainers {
-		if spec.InitContainers[i].Name == "" {
-			spec.InitContainers[i].Name = fmt.Sprintf("%s-%d", pj.ObjectMeta.Name, i)
-		}
-		spec.InitContainers[i].Env = append(spec.InitContainers[i].Env, kubeEnv(env)...)
-	}
-	for i := range spec.Containers {
-		if spec.Containers[i].Name == "" {
-			spec.Containers[i].Name = fmt.Sprintf("%s-%d", pj.ObjectMeta.Name, i)
-		}
-		spec.Containers[i].Env = append(spec.Containers[i].Env, kubeEnv(env)...)
-	}
-	podLabels := make(map[string]string)
-	for k, v := range pj.ObjectMeta.Labels {
-		podLabels[k] = v
-	}
-	podLabels[kube.CreatedByProw] = "true"
-	podLabels[kube.ProwJobTypeLabel] = string(pj.Spec.Type)
-	return &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:   pj.ObjectMeta.Name,
-			Labels: podLabels,
-			Annotations: map[string]string{
-				kube.ProwJobAnnotation: pj.Spec.Job,
-			},
-		},
-		Spec: *spec,
-	}, nil
-}
-
-// kubeEnv transforms a mapping of environment variables
-// into their serialized form for a PodSpec, sorting by
-// the name of the env vars
-func kubeEnv(environment map[string]string) []v1.EnvVar {
-	var keys []string
-	for key := range environment {
-		keys = append(keys, key)
-	}
-	sort.Strings(keys)
-
-	var kubeEnvironment []v1.EnvVar
-	for _, key := range keys {
-		kubeEnvironment = append(kubeEnvironment, v1.EnvVar{
-			Name:  key,
-			Value: environment[key],
-		})
-	}
-
-	return kubeEnvironment
 }
 
 // PartitionActive separates the provided prowjobs into pending and triggered
