+++ prow/pjutil/pjutil_test.go
@@ -20,107 +20,10 @@ import (
 	"testing"
 	"time"
 
-	"k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/equality"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/diff"
 
 	"k8s.io/test-infra/prow/kube"
 )
-
-func TestProwJobToPod(t *testing.T) {
-	tests := []struct {
-		podName string
-		buildID string
-		labels  map[string]string
-		pjSpec  kube.ProwJobSpec
-
-		expected *v1.Pod
-	}{
-		{
-			podName: "pod",
-			buildID: "blabla",
-			labels:  map[string]string{"needstobe": "inherited"},
-			pjSpec: kube.ProwJobSpec{
-				Type:  kube.PresubmitJob,
-				Job:   "job-name",
-				Agent: kube.KubernetesAgent,
-				Refs: &kube.Refs{
-					Org:     "org-name",
-					Repo:    "repo-name",
-					BaseRef: "base-ref",
-					BaseSHA: "base-sha",
-					Pulls: []kube.Pull{{
-						Number: 1,
-						Author: "author-name",
-						SHA:    "pull-sha",
-					}},
-				},
-				PodSpec: &v1.PodSpec{
-					Containers: []v1.Container{
-						{
-							Image: "tester",
-							Env: []v1.EnvVar{
-								{Name: "MY_ENV", Value: "rocks"},
-							},
-						},
-					},
-				},
-			},
-
-			expected: &v1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "pod",
-					Labels: map[string]string{
-						kube.CreatedByProw:    "true",
-						kube.ProwJobTypeLabel: "presubmit",
-						"needstobe":           "inherited",
-					},
-					Annotations: map[string]string{
-						kube.ProwJobAnnotation: "job-name",
-					},
-				},
-				Spec: v1.PodSpec{
-					RestartPolicy: "Never",
-					Containers: []v1.Container{
-						{
-							Name:  "pod-0",
-							Image: "tester",
-							Env: []v1.EnvVar{
-								{Name: "MY_ENV", Value: "rocks"},
-								{Name: "BUILD_ID", Value: "blabla"},
-								{Name: "BUILD_NUMBER", Value: "blabla"},
-								{Name: "JOB_NAME", Value: "job-name"},
-								{Name: "JOB_SPEC", Value: `{"type":"presubmit","job":"job-name","buildid":"blabla","prowjobid":"pod","refs":{"org":"org-name","repo":"repo-name","base_ref":"base-ref","base_sha":"base-sha","pulls":[{"number":1,"author":"author-name","sha":"pull-sha"}]}}`},
-								{Name: "JOB_TYPE", Value: "presubmit"},
-								{Name: "PROW_JOB_ID", Value: "pod"},
-								{Name: "PULL_BASE_REF", Value: "base-ref"},
-								{Name: "PULL_BASE_SHA", Value: "base-sha"},
-								{Name: "PULL_NUMBER", Value: "1"},
-								{Name: "PULL_PULL_SHA", Value: "pull-sha"},
-								{Name: "PULL_REFS", Value: "base-ref:base-sha,1:pull-sha"},
-								{Name: "REPO_NAME", Value: "repo-name"},
-								{Name: "REPO_OWNER", Value: "org-name"},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for i, test := range tests {
-		t.Logf("test run #%d", i)
-		pj := kube.ProwJob{ObjectMeta: metav1.ObjectMeta{Name: test.podName, Labels: test.labels}, Spec: test.pjSpec}
-		got, err := ProwJobToPod(pj, test.buildID)
-		if err != nil {
-			t.Errorf("unexpected error: %v", err)
-		}
-		if !equality.Semantic.DeepEqual(got, test.expected) {
-			t.Errorf("expected pod:\n%s", diff.ObjectReflectDiff(test.expected, got))
-		}
-	}
-}
 
 func TestPartitionActive(t *testing.T) {
 	tests := []struct {
