+++ prow/plank/BUILD.bazel
@@ -31,6 +31,7 @@ go_library(
         "//prow/github:go_default_library",
         "//prow/kube:go_default_library",
         "//prow/pjutil:go_default_library",
+        "//prow/pod-utils/decorate:go_default_library",
         "//prow/report:go_default_library",
         "//vendor/github.com/bwmarrin/snowflake:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
