+++ prow/pod-utils/decorate/OWNERS
@@ -1,0 +1,2 @@
+approvers:
+- stevekuznetsov
