+++ prow/pod-utils/downwardapi/BUILD.bazel
@@ -1,0 +1,33 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "doc.go",
+        "jobspec.go",
+    ],
+    importpath = "k8s.io/test-infra/prow/pod-utils/downwardapi",
+    visibility = ["//visibility:public"],
+    deps = ["//prow/kube:go_default_library"],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["jobspec_test.go"],
+    embed = [":go_default_library"],
+    deps = ["//prow/kube:go_default_library"],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)
