+++ prow/pod-utils/downwardapi/OWNERS
@@ -1,0 +1,2 @@
+approvers:
+- stevekuznetsov
