+++ prow/pod-utils/downwardapi/jobspec_test.go
@@ -14,12 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package pjutil
+package downwardapi
 
 import (
-	"fmt"
-	"net/http"
-	"net/http/httptest"
 	"reflect"
 	"testing"
 
@@ -191,84 +188,5 @@ func TestEnvironmentForSpec(t *testing.T) {
 		if actual, expected := env, test.expected; !reflect.DeepEqual(actual, expected) {
 			t.Errorf("%s: got environment:\n\t%v\n\tbut expected:\n\t%v", test.name, actual, expected)
 		}
-	}
-}
-
-type responseVendor struct {
-	codes []int
-	data  []string
-
-	position int
-}
-
-func (r *responseVendor) next() (int, string) {
-	code := r.codes[r.position]
-	datum := r.data[r.position]
-
-	r.position = r.position + 1
-	if r.position == len(r.codes) {
-		r.position = 0
-	}
-
-	return code, datum
-}
-
-func parrotServer(codes []int, data []string) *httptest.Server {
-	vendor := responseVendor{
-		codes: codes,
-		data:  data,
-	}
-
-	return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		code, datum := vendor.next()
-		w.WriteHeader(code)
-		fmt.Fprint(w, datum)
-	}))
-}
-
-func TestGetBuildID(t *testing.T) {
-	var testCases = []struct {
-		name        string
-		codes       []int
-		data        []string
-		expected    string
-		expectedErr bool
-	}{
-		{
-			name:        "all good",
-			codes:       []int{200},
-			data:        []string{"yay"},
-			expected:    "yay",
-			expectedErr: false,
-		},
-		{
-			name:        "fail then success",
-			codes:       []int{500, 200},
-			data:        []string{"boo", "yay"},
-			expected:    "yay",
-			expectedErr: false,
-		},
-		{
-			name:        "fail",
-			codes:       []int{500},
-			data:        []string{"boo"},
-			expected:    "boo",
-			expectedErr: true,
-		},
-	}
-
-	for _, testCase := range testCases {
-		totServ := parrotServer(testCase.codes, testCase.data)
-
-		actual, actualErr := GetBuildID("dummy", totServ.URL)
-		if testCase.expectedErr && actualErr == nil {
-			t.Errorf("%s: expected an error but got none", testCase.name)
-		} else if !testCase.expectedErr && actualErr != nil {
-			t.Errorf("%s: expected no error but got one: %v", testCase.name, actualErr)
-		} else if !testCase.expectedErr && actual != testCase.expected {
-			t.Errorf("%s: expected response %v but got: %v", testCase.name, testCase.expected, actual)
-		}
-
-		totServ.Close()
 	}
 }
