+++ prow/pod-utils/gcs/options.go
@@ -26,10 +26,11 @@ import (
 	"path/filepath"
 	"strings"
 
+	"k8s.io/test-infra/prow/pod-utils/downwardapi"
+
 	"cloud.google.com/go/storage"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/api/option"
-	"k8s.io/test-infra/prow/pjutil"
 )
 
 type pathStrategyType string
@@ -123,7 +124,7 @@ func (o *Options) Run(extra map[string]UploadFunc) error {
 		builder = NewSingleDefaultRepoPathBuilder(o.DefaultOrg, o.DefaultRepo)
 	}
 
-	spec, err := pjutil.ResolveSpecFromEnv()
+	spec, err := downwardapi.ResolveSpecFromEnv()
 	if err != nil {
 		return fmt.Errorf("could not resolve job spec: %v", err)
 	}
